pyproject.toml
[tool.poetry]
name = "netrunner-db"
version = "0.1.0"
description = "export Netrunner card data to ElasticSearch."
authors = ["Eric Li <eric03742@foxmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.0"
meilisearch = "^0.31.6"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
